refactor: clarify the action and targets
I think that a function and class should represent an action and a target clearly for clean code.

So, I renamed them as follows:

1. class Painting to ImageColorsSimplifier
Painting word can’t represents anything.
according to my analysis, Painting class has three process, which are blurring, upscaling, and clustering.
and the blurring and upscaling processes exist to simplify the image colors, and upscaling should be processed before clustering because it expands image by Interpolation. (Interpolation increase num of colors)
So I think it should be renamed to “ImageColorSimpifier”.

2. blurring() -> blur_image()
- In the case of blurring, it's not a verb, and there is no target.

3. cluster_color_with_kmeans() -> cluster_image_color()
- The k-means should be removed as it is just algorithm to cluster colors, not a behavioral name.

4. LineDrawer -> ColorSectorLineDrawer
Where does it draw the line? The target is unclear.
“ColorSectorLineDrawer” is probably right, because it draws lines in sectors divided by color.